requirements.txt
Django==1.10.3
django-autocomplete-light==3.2.1
django-debug-toolbar==1.6
django-suit==0.2.22
Pillow==3.4.2
psycopg2==2.6.2
reportlab==3.3.0
sqlparse==0.2.2
